refactor RatingFragment, remove nav graph, refactor style to comply accessibility
