+++ bandyer-android-design/src/main/res/layout/bandyer_rating_dialog_layout.xml
@@ -1,17 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     xmlns:tools="http://schemas.android.com/tools"
+    android:animateLayoutChanges="true"
     tools:context="rating.RatingDialog">
 
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/mainNavFragment"
-        android:name="androidx.navigation.fragment.NavHostFragment"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:defaultNavHost="false"
-        app:navGraph="@navigation/bandyer_rating_dialog_graph" />
+    <include
+        android:id="@+id/bandyer_rating_layout"
+        layout="@layout/fragment_rating_layout"/>
 
-</FrameLayout>
+    <include
+        android:id="@+id/bandyer_rating_done_layout"
+        android:visibility="gone"
+        layout="@layout/fragment_rating_done_layout"/>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
