+++ bandyer-android-design/src/main/res/layout/fragment_rating_done_layout.xml
@@ -5,8 +5,7 @@
     android:layout_height="wrap_content"
     android:gravity="center"
     android:orientation="vertical"
-    android:padding="@dimen/bandyer_dimen_space16"
-    tools:context=".rating.RateAfterFragment">
+    android:padding="@dimen/bandyer_dimen_space16">
 
     <com.google.android.material.textview.MaterialTextView
         android:id="@+id/bandyer_title"
@@ -29,13 +28,13 @@
         android:id="@+id/bandyer_subtitle"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:gravity="center"
         android:layout_marginStart="@dimen/bandyer_dimen_space4"
+        android:layout_marginTop="@dimen/bandyer_dimen_space16"
         android:layout_marginEnd="@dimen/bandyer_dimen_space4"
+        android:gravity="center"
         android:lineSpacingExtra="2dp"
         android:paddingStart="@dimen/bandyer_dimen_space4"
         android:paddingEnd="@dimen/bandyer_dimen_space4"
-        android:layout_marginTop="@dimen/bandyer_dimen_space16"
         android:text="Speriamo di rivederti presto."
         android:textColor="@color/bandyer_color_on_surface_50"
         android:textSize="16sp"
