+++ bandyer-android-design/src/main/res/layout/fragment_rating_layout.xml
@@ -1,20 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:padding="@dimen/bandyer_dimen_space4"
-    tools:context=".rating.RateFragment">
+    android:padding="@dimen/bandyer_dimen_space4">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/bandyer_close"
+        style="@style/BandyerSDKDesign.ImageButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         app:icon="@drawable/ic_bandyer_close"
+        app:iconSize="@dimen/bandyer_dimen_space24"
         app:iconTint="?attr/colorOnSurface"
-        app:iconSize="@dimen/bandyer_dimen_space20"
-        style="@style/BandyerSDKDesign.ImageButton"
+        android:contentDescription="close"
+        android:paddingTop="@dimen/bandyer_dimen_space12"
+        android:paddingStart="@dimen/bandyer_dimen_space12"
+        android:paddingEnd="@dimen/bandyer_dimen_space12"
+        android:paddingBottom="@dimen/bandyer_dimen_space12"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -27,8 +32,9 @@
         android:layout_marginEnd="@dimen/bandyer_dimen_space28"
         android:gravity="center"
         android:lineSpacingExtra="2dp"
+        android:contentDescription="Valuta la qualità della video chiamata"
         android:text="Valuta la qualità della video chiamata"
-        android:textColor="@color/bandyer_color_on_surface_50"
+        android:textColor="@color/bandyer_color_on_surface_80"
         android:textSize="16sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -40,14 +46,17 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/bandyer_dimen_space28"
-        android:layout_marginTop="@dimen/bandyer_dimen_space24"
+        android:layout_marginTop="@dimen/bandyer_dimen_space13"
         android:layout_marginEnd="@dimen/bandyer_dimen_space28"
         android:clickable="true"
         android:focusable="true"
-        android:rating="1"
+        android:rating="5"
         app:bandyer_minRating="1"
         app:bandyer_numLevels="5"
-        app:drawableSize="@dimen/bandyer_dimen_space24"
+        android:paddingTop="@dimen/bandyer_dimen_space10"
+        android:paddingBottom="@dimen/bandyer_dimen_space10"
+        android:contentDescription="Rating bar"
+        app:drawableSize="@dimen/bandyer_dimen_space28"
         app:drawableTint="?attr/colorSecondary"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -58,7 +67,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/bandyer_dimen_space28"
-        android:layout_marginTop="@dimen/bandyer_dimen_space8"
         android:layout_marginEnd="@dimen/bandyer_dimen_space28"
         android:gravity="center"
         android:letterSpacing="0.03"
@@ -66,6 +74,7 @@
         android:textColor="?attr/colorOnSurface"
         android:textSize="16sp"
         android:textStyle="bold"
+        android:contentDescription="Molto male"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/bandyer_rating"
@@ -94,16 +103,11 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:maxLines="4"
-            android:textSize="14sp"
             android:scrollbars="vertical"
             android:textAlignment="viewStart"
             android:textDirection="locale"
-            android:scrollbarStyle="insideInset"
+            android:scrollbarStyle="outsideInset"
             android:scrollbarThumbVertical="@drawable/bandyer_rating_scrollbar_thumb"
-            android:paddingTop="@dimen/bandyer_dimen_space8"
-            android:paddingBottom="@dimen/bandyer_dimen_space8"
-            android:paddingStart="@dimen/bandyer_dimen_space8"
-            android:paddingEnd="@dimen/bandyer_dimen_space8"
             android:textColorHint="@color/bandyer_color_on_surface_50"
             android:textAppearance="@style/BandyerSDKDesign.TextAppearance.TextInputEditText"
             android:hint="Lascia un commento..."
