--- bandyer-android-design/src/main/res/navigation/bandyer_rating_dialog_graph.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/bandyer_rating_dialog_graph"
-    app:startDestination="@id/rateFragment">
-
-    <fragment
-        android:id="@+id/rateFragment"
-        android:name="com.bandyer.sdk_design.rating.RateFragment"
-        android:label="fragment_rate"
-        tools:layout="@layout/fragment_rate" >
-        <action
-            android:id="@+id/action_rateFragment_to_rateAfterFragment"
-            app:destination="@id/rateAfterFragment"
-            app:popUpTo="@id/rateAfterFragment"
-            app:popUpToInclusive="true" />
-    </fragment>
-    <fragment
-        android:id="@+id/rateAfterFragment"
-        android:name="com.bandyer.sdk_design.rating.RateAfterFragment"
-        android:label="fragment_rate_after"
-        tools:layout="@layout/fragment_rate_after" />
-</navigation>
