+++ bandyer-android-design/src/main/res/values/text_input.xml
@@ -43,7 +43,6 @@
     <style name="BandyerSDKDesign.TextInputEditText" parent="Base.Widget.MaterialComponents.TextInputEditText">
         <item name="android:background">@null</item>
         <item name="android:gravity">start</item>
-        <item name="android:textColor">@null</item>
         <item name="android:hint">@string/bandyer_edit_text_input_placeholder</item>
         <item name="android:textColorHint">@color/bandyer_color_on_surface_50</item>
         <item name="android:textAlignment" tools:targetApi="jelly_bean_mr1">viewStart</item>
@@ -55,6 +54,7 @@
         <item name="android:shadowDy">0</item>
         <item name="android:shadowRadius">0</item>
         <item name="android:textColor">?attr/colorOnSurface</item>
+<!--        <item name="android:textSize">14sp</item>-->
     </style>
 
     <style name="BandyerSDKDesign.TextInputLayout.Chat" parent="BandyerSDKDesign.TextInputLayout">
